flask_app/requirements.txt
gunicorn
annotated-types==0.7.0
async-timeout==4.0.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
docstring-parser==0.16
eventlet==0.36.1
flask==3.0.3
flask-sock==0.7.0
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
idna==3.7
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.8.2
pydantic-core==2.20.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
redis==5.0.8
requests==2.32.3
rsa==4.9
schedule==1.2.2
shapely==2.0.5
simple-websocket==1.0.0
six==1.16.0
typing-extensions==4.12.2
urllib3==2.2.2
vertexai==1.60.0
websocket==0.2.1
websocket-client==1.8.0
werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
zope.event==5.0
zope.interface==7.0.1
